Move spp integration to libr.asm.c
also remove the spp parts from rasm2.c
Fix #6356
