+++ binr/rasm2/Makefile
@@ -1,5 +1,4 @@
 BIN=rasm2
 BINDEPS=r_asm r_parse r_syscall r_anal r_reg
 BINDEPS+=r_flag r_cons r_lang r_util
-CFLAGS+=-I$(SHLR)/spp
 include ../rules.mk
